Move contract detail coverage tab to use common adapter
