+++ app/src/main/java/com/hedvig/app/feature/insurablelimits/InsurableLimitsAdapter.kt
@@ -87,15 +87,3 @@ class InsurableLimitsAdapter(
     }
 }
 
-sealed class InsurableLimitItem {
-    sealed class Header : InsurableLimitItem() {
-        object Details : Header()
-        object MoreInfo : Header()
-    }
-
-    data class InsurableLimit(
-        val label: String,
-        val limit: String,
-        val description: String,
-    ) : InsurableLimitItem()
-}
