+++ app/src/main/java/com/hedvig/app/feature/insurance/ui/detail/ContractDetailViewModel.kt
@@ -11,7 +11,9 @@ import com.hedvig.app.R
 import com.hedvig.app.feature.chat.data.ChatRepository
 import com.hedvig.app.feature.documents.DocumentItems
 import com.hedvig.app.feature.insurance.data.InsuranceRepository
-import com.hedvig.app.feature.insurance.ui.detail.coverage.CoverageModel
+import com.hedvig.app.feature.insurance.ui.detail.coverage.CoverageViewState
+import com.hedvig.app.feature.insurance.ui.detail.coverage.createCoverageItems
+import com.hedvig.app.feature.insurance.ui.detail.coverage.createInsurableLimitsItems
 import com.hedvig.app.feature.insurance.ui.detail.yourinfo.YourInfoModel
 import com.hedvig.app.feature.settings.Market
 import com.hedvig.app.feature.settings.MarketManager
@@ -30,8 +32,8 @@ abstract class ContractDetailViewModel : ViewModel() {
     protected val _documentsList = MutableLiveData<List<DocumentItems.Document>>()
     val documentsList: LiveData<List<DocumentItems.Document>> = _documentsList
 
-    protected val _coverageList = MutableLiveData<List<CoverageModel>>()
-    val coverageList: LiveData<List<CoverageModel>> = _coverageList
+    protected val _coverageViewState = MutableLiveData<CoverageViewState>()
+    val coverageViewState: LiveData<CoverageViewState> = _coverageViewState
 
     abstract fun loadContract(id: String)
     abstract suspend fun triggerFreeTextChat()
@@ -40,7 +42,7 @@ abstract class ContractDetailViewModel : ViewModel() {
 class ContractDetailViewModelImpl(
     private val insuranceRepository: InsuranceRepository,
     private val chatRepository: ChatRepository,
-    private val marketManager: MarketManager
+    private val marketManager: MarketManager,
 ) : ContractDetailViewModel() {
 
     override fun loadContract(id: String) {
@@ -56,7 +58,12 @@ class ContractDetailViewModelImpl(
                             _data.postValue(Result.success(contract))
                             _yourInfoList.postValue(createContractItems(contract))
                             _documentsList.postValue(createDocumentItems(contract))
-                            _coverageList.postValue(createCoverageItems(contract))
+                            _coverageViewState.postValue(
+                                CoverageViewState(
+                                    createCoverageItems(contract),
+                                    createInsurableLimitsItems(contract),
+                                )
+                            )
                         } ?: _data.postValue(Result.failure(Throwable("No contract found")))
                 }
             }
@@ -100,16 +107,6 @@ class ContractDetailViewModelImpl(
                     )
                 }
             )
-        }
-    }
-
-    private fun createCoverageItems(contract: InsuranceQuery.Contract): List<CoverageModel> {
-        return listOf(
-            CoverageModel.Header.Perils(contract.typeOfContract)
-        ) + contract.perils.map {
-            CoverageModel.Peril(it.fragments.perilFragment)
-        } + CoverageModel.Header.InsurableLimits + contract.insurableLimits.map {
-            CoverageModel.InsurableLimit(it.fragments.insurableLimitsFragment)
         }
     }
