+++ app/src/main/res/layout/contract_detail_coverage_header.xml
@@ -1,9 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <TextView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_marginHorizontal="@dimen/base_margin_double"
     android:layout_marginTop="@dimen/base_margin_quintuple"
     android:layout_marginBottom="@dimen/base_margin"
-    tools:text="@string/CONTRACT_COVERAGE_CONTRACT_TYPE"
-    android:textAppearance="?textAppearanceHeadline6" />
+    android:textAppearance="?textAppearanceHeadline6"
+    tools:text="@string/CONTRACT_COVERAGE_CONTRACT_TYPE" />
